Update digital_score_reel.py
This PR will pull the left-most digits from the player's score and display them on the digital score reels.  The original code pulled the right-most digits so if a score overflow occurred a five digit score reel would display 123,000 as 12,300 not 23,000.  With this change the user can make a traditional EM overflow light widget to display 100,000 along with the five digital score reels.